# last_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MusicAI.Orchestrator",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Radio/stream/{trackId}": {
      "get": {
        "tags": [
          "Radio"
        ],
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Radio/auto-play/next": {
      "get": {
        "tags": [
          "Radio"
        ],
        "parameters": [
          {
            "name": "lastTrackId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}